Implemented option to disable postgis

+++ migrations/20160428234902-client_location_index.js
@@ -1,11 +1,20 @@
 'use strict';
+var q = require("q");
 
 module.exports = {
   up: function (queryInterface, Sequelize) {
+    if (process.env.POSTGIS_DISABLED) {
+      return q();
+    }
+
     return queryInterface.sequelize.query("CREATE INDEX client_location_index ON clients USING GIST (location)");
   },
 
   down: function (queryInterface, Sequelize) {
+    if (process.env.POSTGIS_DISABLED) {
+      return q();
+    }
+
     return queryInterface.sequelize.query("DROP INDEX client_location_index");
   }
 };
